backend/requirements.txt
flask>=3.0,<4
flask-cors>=4.0,<5
python-dotenv>=1.0,<2
pymongo>=4.6,<5
bcrypt>=4.1,<5
PyJWT>=2.8,<3
pillow>=10.3,<11
numpy>=1.26,<3

# Emotion inference (install TensorFlow build appropriate for your OS/Python)
tensorflow>=2.15,<3

# Optional: face verification (requires CMake + dlib on many systems; see README)
# Uncomment on Linux/WSL/macOS when ready:
# face_recognition>=1.3.0
